requirements.txt
pandas
streamlit
langchain
os
OpenAI
python-dotenv
Pillow
streamlit-extras
streamlit_chat
sentence_transformers
pinecone-client
